Permanent operation of IRKit without a server
When infrared signal (IR) is received, the LED flashes green-->red-->blue-->turns off.
It seems that the software of Wifi is reset by IR reception.
In this state, the IR signal cannot be learned, and it is dedicated to IR transmission.
By rewriting this new firmware, IR transmission and reception can be operated permanently without a server.